% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bareme}
\alias{bareme}
\title{Scale use to compute ranking}
\format{
A data frame with 22 rows and 9 columns
\describe{
\item{classement}{Ranking to test}
\item{index_min}{Lower boundary of the "V-E-2I-5G" index giving the number of
matches to add in points calculation}
\item{index_max}{Higher boundary of the "V-E-2I-5G" index giving the number of
matches to add in points calculation}
\item{match_en_plus}{Number of matches to add to the number of matches to
take into account in the point calculation step}
\item{serie}{division of ranking ("4","3","2plus","2moins")}
\item{point_min_h}{Minimum points required to validate ranking for men}
\item{nb_victoire_compte_h}{Fixed number of matches took into account while
calculating points for men}
\item{point_min_f}{Minimum points required to validate ranking for women}
\item{nb_victoire_compte_f}{Fixed number of matches took into account while
calculating points for men}
}
}
\usage{
bareme
}
\description{
This dataset gather all necessary parameters to compute ranking of french
tennis player playing in FFT (French Federation of Tennis). This data frame
is available on the \href{https://tenup.fft.fr/content/classement-tennis-bareme-et-normes}{FFT website}
}
\references{
\url{https://tenup.fft.fr/}
}
\keyword{datasets}
