% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_to_df.R
\name{html_to_df}
\alias{html_to_df}
\title{Convert html palmares extraction into a data frame}
\usage{
html_to_df(file)
}
\arguments{
\item{file}{html file to convert}
}
\value{
data.frame of match records in adequate format for ranking simulation
}
\description{
After extracting a palmares of tennis matches on \href{https://tenup.fft.fr}{tenup}
(to do so, see \link{tutorial}), this function allows to convert your html match
records into a data.frame
}
\examples{
library(tenupR)
file <- list.files(system.file("extdata", package = "tenupR"),full.names = T)
df <- html_to_df(file)
head(df)
}
