% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{palmares}
\alias{palmares}
\title{Example of record data from a player needed to compute ranking}
\format{
A data frame with 23 rows and 10 columns
\describe{
\item{adversaire}{Name of the opponent}
\item{ne_en}{Date of birth of the opponent}
\item{classement}{Opponent ranking}
\item{issue}{Outcome of the match, stated as "V" for wins and "D" for loses}
\item{score}{Final score of the match (WO)}
\item{competition}{Context of the match (team cup,TMC, single championship)}
\item{coef}{coeficient applied on the match}
\item{joue_le}{Date of the match}
\item{date_val}{Date until the match will be counted in the ranking process}
\item{match_type}{type of match(single or double)}
}
}
\usage{
palmares
}
\description{
Record of player matches are available for each french player
on \url{https://tenup.fft.fr/} requiring a tenup account.
All of these information are used to compute ranking of a player.
}
\references{
\url{https://tenup.fft.fr/}
}
\keyword{datasets}
