% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/point.R
\name{points}
\alias{points}
\title{Get amount of points from a difference of ranking}
\usage{
points(diff)
}
\arguments{
\item{diff}{difference of rankings as a negative or positive integer.}
}
\value{
number of points earned from the difference of ranking
}
\description{
From a difference of rankings (positive or negative) get the amount of
points won. The amount of points is given by a scale set by FFT that is
available \href{https://tenup.fft.fr/content/classement-tennis-bareme-et-normes}{here}
}
\examples{
library(tenupR)
# How many points do I earn if my ranking is 15/5 and I win against a 15/3 ?
my_ranking <- as_ranking("15/5")
his_ranking <- as_ranking("15/3")
diff <- as.numeric(his_ranking) - as.numeric(my_ranking)
points(diff)
}
\seealso{
\link{as_ranking}
}
